refactor: Rename CreateSwearJarForm to SwearJarForm and /component/app/swearjar/create folder to /component/app/swearjar/forms
