+++ frontend/app/swearjar/create/mainContent.tsx
@@ -1,4 +1,4 @@
-import CreateSwearJarForm from "@/components/app/swearjar/create/CreateSwearJarForm";
+import CreateSwearJarForm from "@/components/app/swearjar/forms/SwearJarForm";
 
 export default function MainContent() {
     return (
